refactor: split shared types and demo data by domain
