+++ packages/shared/src/data/ai-logs.ts
@@ -1,0 +1,197 @@
+import type { AIIntentLog, AIQuestionLog, AIUsageLog } from '../types';
+
+export let demoAIUsageLogs: AIUsageLog[] = [
+  {
+    id: 'aiu_001',
+    user_id: 'usr_std_001',
+    feature: 'smart_chat',
+    provider: 'workers_ai',
+    model: 'llama-3.1-70b',
+    input_tokens: 156,
+    output_tokens: 284,
+    latency_ms: 820,
+    success: 1,
+    error_message: null,
+    created_at: '2026-04-08T08:00:00.000Z',
+  },
+  {
+    id: 'aiu_002',
+    user_id: 'usr_std_001',
+    feature: 'summary',
+    provider: 'workers_ai',
+    model: 'llama-3.1-70b',
+    input_tokens: 192,
+    output_tokens: 220,
+    latency_ms: 740,
+    success: 1,
+    error_message: null,
+    created_at: '2026-04-08T08:05:00.000Z',
+  },
+  {
+    id: 'aiu_003',
+    user_id: 'usr_std_001',
+    feature: 'quiz',
+    provider: 'workers_ai',
+    model: 'llama-3.1-70b',
+    input_tokens: 205,
+    output_tokens: 310,
+    latency_ms: 960,
+    success: 1,
+    error_message: null,
+    created_at: '2026-04-08T08:12:00.000Z',
+  },
+  {
+    id: 'aiu_004',
+    user_id: 'usr_inst_001',
+    feature: 'qa',
+    provider: 'workers_ai',
+    model: 'llama-3.1-70b',
+    input_tokens: 240,
+    output_tokens: 390,
+    latency_ms: 1120,
+    success: 1,
+    error_message: null,
+    created_at: '2026-04-08T09:00:00.000Z',
+  },
+  {
+    id: 'aiu_005',
+    user_id: 'usr_inst_001',
+    feature: 'qa',
+    provider: 'workers_ai',
+    model: 'llama-3.1-70b',
+    input_tokens: 220,
+    output_tokens: 260,
+    latency_ms: 1300,
+    success: 0,
+    error_message: 'context_not_found',
+    created_at: '2026-04-08T09:10:00.000Z',
+  },
+  {
+    id: 'aiu_006',
+    user_id: 'usr_admin_001',
+    feature: 'insights',
+    provider: 'workers_ai',
+    model: 'llama-3.1-70b',
+    input_tokens: 95,
+    output_tokens: 180,
+    latency_ms: 620,
+    success: 1,
+    error_message: null,
+    created_at: '2026-04-08T09:20:00.000Z',
+  },
+];
+
+export let demoAIIntentLogs: AIIntentLog[] = [
+  {
+    id: 'aint_001',
+    user_id: 'usr_std_001',
+    message: '이 강의 핵심만 요약해줘',
+    detected_intent: 'request_summary',
+    confidence: 0.98,
+    feature: 'summary',
+    success: true,
+    action_taken: 'summarized',
+    lecture_id: 'lec_llm_001',
+    course_id: 'crs_llm_001',
+    created_at: '2026-04-08T08:05:00.000Z',
+  },
+  {
+    id: 'aint_002',
+    user_id: 'usr_std_001',
+    message: '퀴즈 만들어줘',
+    detected_intent: 'generate_quiz',
+    confidence: 0.92,
+    feature: 'quiz',
+    success: true,
+    action_taken: 'quiz_generated',
+    lecture_id: 'lec_llm_002',
+    course_id: 'crs_llm_001',
+    created_at: '2026-04-08T08:12:00.000Z',
+  },
+  {
+    id: 'aint_003',
+    user_id: 'usr_std_001',
+    message: '이 개념을 다시 설명해줘',
+    detected_intent: 'ask_concept',
+    confidence: 0.95,
+    feature: 'smart_chat',
+    success: true,
+    action_taken: 'rag_answered',
+    lecture_id: 'lec_llm_002',
+    course_id: 'crs_llm_001',
+    created_at: '2026-04-08T08:20:00.000Z',
+  },
+  {
+    id: 'aint_004',
+    user_id: 'usr_inst_001',
+    message: '우리 강의에서 질문이 가장 많이 나온 주제가 뭐야?',
+    detected_intent: 'analyze_progress',
+    confidence: 0.88,
+    feature: 'insights',
+    success: true,
+    action_taken: 'instructor_insight',
+    lecture_id: null,
+    course_id: 'crs_llm_001',
+    created_at: '2026-04-08T09:00:00.000Z',
+  },
+  {
+    id: 'aint_005',
+    user_id: 'usr_admin_001',
+    message: '이번 주 AI 사용량을 보여줘',
+    detected_intent: 'analyze_progress',
+    confidence: 0.93,
+    feature: 'insights',
+    success: true,
+    action_taken: 'admin_overview',
+    lecture_id: null,
+    course_id: null,
+    created_at: '2026-04-08T09:20:00.000Z',
+  },
+];
+
+export let demoAIQuestionLogs: AIQuestionLog[] = [
+  {
+    id: 'aiq_001',
+    user_id: 'usr_std_001',
+    lecture_id: 'lec_llm_001',
+    course_id: 'crs_llm_001',
+    question: 'RAG는 왜 청킹이 필요한가요?',
+    answer: '청킹은 검색 단위를 잘게 나눠 관련 근거를 찾기 쉽게 해 줍니다.',
+    model: 'llama-3.1-70b',
+    success: true,
+    created_at: '2026-04-08T08:21:00.000Z',
+  },
+  {
+    id: 'aiq_002',
+    user_id: 'usr_std_001',
+    lecture_id: 'lec_llm_002',
+    course_id: 'crs_llm_001',
+    question: '요약과 질문응답을 어떻게 구분해야 하나요?',
+    answer: '요약은 원문 구조를 압축하고, 질문응답은 특정 질문에 근거를 붙여 답합니다.',
+    model: 'llama-3.1-70b',
+    success: true,
+    created_at: '2026-04-08T08:25:00.000Z',
+  },
+  {
+    id: 'aiq_003',
+    user_id: 'usr_inst_001',
+    lecture_id: 'lec_llm_001',
+    course_id: 'crs_llm_001',
+    question: 'RAG 파이프라인에서 가장 중요한 지점은?',
+    answer: '근거를 잃지 않는 검색과 재조합입니다.',
+    model: 'llama-3.1-70b',
+    success: true,
+    created_at: '2026-04-08T09:02:00.000Z',
+  },
+  {
+    id: 'aiq_004',
+    user_id: 'usr_inst_001',
+    lecture_id: 'lec_llm_002',
+    course_id: 'crs_llm_001',
+    question: '학생들이 제일 많이 묻는 개념은?',
+    answer: '청킹과 근거 검색에 대한 질문이 많습니다.',
+    model: 'llama-3.1-70b',
+    success: true,
+    created_at: '2026-04-08T09:06:00.000Z',
+  },
+];
