+++ packages/shared/src/data/demo-data.ts
@@ -1,436 +1,4 @@
-import type {
-  AudioExtraction,
-  AIIntentLog,
-  AIQuestionLog,
-  AIUsageLog,
-  AuthUser,
-  Course,
-  Enrollment,
-  Lecture,
-  LectureNote,
-  LecturePipeline,
-  LectureProgress,
-  LectureTranscript,
-  Material,
-  Notice,
-} from '../types';
-
-export const demoUsers: AuthUser[] = [
-  {
-    id: 'usr_std_001',
-    name: '데모 수강생',
-    email: 'student@mywayclass.local',
-    role: 'STUDENT',
-    department: '학습자',
-    bio: '강의 숏폼과 커스텀 강의를 활용해 복습 효율을 높이고 싶은 수강생입니다.',
-  },
-  {
-    id: 'usr_inst_001',
-    name: '김민준',
-    email: 'instructor1@mywayclass.local',
-    role: 'INSTRUCTOR',
-    department: 'AI 교육학부',
-    bio: '학습 데이터를 기반으로 더 나은 강의를 설계하는 강사입니다.',
-  },
-  {
-    id: 'usr_admin_001',
-    name: '오운영',
-    email: 'admin@mywayclass.local',
-    role: 'ADMIN',
-    department: '교육 운영팀',
-    bio: '수강 흐름과 운영 데이터를 관리하는 운영자입니다.',
-  },
-];
-
-export function getDemoUser(userId: string): AuthUser | undefined {
-  return demoUsers.find((user) => user.id === userId);
-}
-
-export const instructorNames: Record<string, string> = {
-  usr_inst_001: '김민준',
-  usr_inst_002: '이서연',
-};
-
-export const demoCourses: Course[] = [
-  {
-    id: 'crs_ai_001',
-    instructor_id: 'usr_inst_001',
-    title: 'AI 기초와 학습 데이터 이해',
-    description: 'AI의 기본 개념과 학습 데이터 구조를 함께 익히는 입문 과정입니다.',
-    category: 'AI',
-    difficulty: 'beginner',
-    is_published: true,
-    tags: ['AI', '기초', '데이터'],
-  },
-  {
-    id: 'crs_web_001',
-    instructor_id: 'usr_inst_002',
-    title: 'React와 TypeScript로 만드는 실전 웹앱',
-    description: '화면 구성, 상태 관리, API 연결을 다루는 웹 개발 과정입니다.',
-    category: 'Web',
-    difficulty: 'intermediate',
-    is_published: true,
-    tags: ['React', 'TypeScript', 'Web'],
-  },
-  {
-    id: 'crs_llm_001',
-    instructor_id: 'usr_inst_001',
-    title: 'RAG와 프롬프트로 만드는 학습 도우미',
-    description: '강의 검색, 요약, 질문응답을 연결하는 AI 학습 보조 과정을 다룹니다.',
-    category: 'AI',
-    difficulty: 'advanced',
-    is_published: true,
-    tags: ['RAG', 'LLM', '프롬프트'],
-  },
-];
-
-export const demoLectures: Lecture[] = [
-  {
-    id: 'lec_ai_001',
-    course_id: 'crs_ai_001',
-    title: 'AI란 무엇인가',
-    order_index: 0,
-    content_type: 'video',
-    content_text: 'AI의 정의와 역사, 머신러닝과 딥러닝의 관계를 살펴봅니다.',
-    duration_minutes: 18,
-    is_published: true,
-  },
-  {
-    id: 'lec_ai_002',
-    course_id: 'crs_ai_001',
-    title: '학습 데이터와 모델의 관계',
-    order_index: 1,
-    content_type: 'video',
-    content_text: '데이터셋 구성, 라벨링, 검증 데이터의 역할을 다룹니다.',
-    duration_minutes: 22,
-    is_published: true,
-  },
-  {
-    id: 'lec_web_001',
-    course_id: 'crs_web_001',
-    title: 'React 컴포넌트 구조',
-    order_index: 0,
-    content_type: 'video',
-    content_text: '함수형 컴포넌트와 상태 흐름, props 전달을 설명합니다.',
-    duration_minutes: 25,
-    is_published: true,
-  },
-  {
-    id: 'lec_web_002',
-    course_id: 'crs_web_001',
-    title: 'TypeScript와 폼 처리',
-    order_index: 1,
-    content_type: 'video',
-    content_text: '폼 검증과 타입 안정성을 함께 유지하는 방법을 소개합니다.',
-    duration_minutes: 24,
-    is_published: true,
-  },
-  {
-    id: 'lec_llm_001',
-    course_id: 'crs_llm_001',
-    title: 'RAG 파이프라인 설계',
-    order_index: 0,
-    content_type: 'video',
-    content_text: '청킹, 임베딩, 검색, 생성의 연결 구조를 설명합니다.',
-    duration_minutes: 30,
-    is_published: true,
-  },
-  {
-    id: 'lec_llm_002',
-    course_id: 'crs_llm_001',
-    title: '강의 요약과 질문응답 연결',
-    order_index: 1,
-    content_type: 'video',
-    content_text: '강의 요약과 챗봇 응답을 하나의 학습 흐름으로 묶는 방법을 다룹니다.',
-    duration_minutes: 28,
-    is_published: true,
-  },
-];
-
-export let demoEnrollments: Enrollment[] = [
-  {
-    id: 'enr_001',
-    user_id: 'usr_std_001',
-    course_id: 'crs_ai_001',
-    status: 'active',
-    progress_percent: 50,
-  },
-  {
-    id: 'enr_002',
-    user_id: 'usr_std_001',
-    course_id: 'crs_llm_001',
-    status: 'active',
-    progress_percent: 20,
-  },
-];
-
-export let demoLectureProgress: LectureProgress[] = [
-  { id: 'prg_001', user_id: 'usr_std_001', lecture_id: 'lec_ai_001', is_completed: true },
-  { id: 'prg_002', user_id: 'usr_std_001', lecture_id: 'lec_ai_002', is_completed: false },
-  { id: 'prg_003', user_id: 'usr_std_001', lecture_id: 'lec_llm_001', is_completed: false },
-  { id: 'prg_004', user_id: 'usr_std_001', lecture_id: 'lec_llm_002', is_completed: false },
-];
-
-export let demoMaterials: Material[] = [
-  {
-    id: 'mat_001',
-    course_id: 'crs_ai_001',
-    title: 'AI 기초 주차 자료',
-    summary: '강의 개요, 핵심 개념, 예제 정리본입니다.',
-    file_name: 'ai-basics-week1.pdf',
-    uploaded_by: 'usr_inst_001',
-    uploaded_at: '2026-04-01T09:00:00.000Z',
-  },
-  {
-    id: 'mat_002',
-    course_id: 'crs_web_001',
-    title: 'React 실습 체크리스트',
-    summary: '컴포넌트, 상태 관리, API 연동 점검표입니다.',
-    file_name: 'react-practice-checklist.pdf',
-    uploaded_by: 'usr_inst_002',
-    uploaded_at: '2026-04-02T09:00:00.000Z',
-  },
-  {
-    id: 'mat_003',
-    course_id: 'crs_llm_001',
-    title: 'RAG 설계 참고자료',
-    summary: '검색, 임베딩, 생성 파이프라인 참고 문서입니다.',
-    file_name: 'rag-design-notes.pdf',
-    uploaded_by: 'usr_inst_001',
-    uploaded_at: '2026-04-03T09:00:00.000Z',
-  },
-];
-
-export let demoNotices: Notice[] = [
-  {
-    id: 'ntc_001',
-    course_id: 'crs_ai_001',
-    title: '1주차 과제 안내',
-    content: '다음 주차 시작 전까지 AI 개념 요약을 제출해주세요.',
-    pinned: true,
-    author_id: 'usr_inst_001',
-    created_at: '2026-04-01T12:00:00.000Z',
-  },
-  {
-    id: 'ntc_002',
-    course_id: 'crs_web_001',
-    title: '실습 환경 점검',
-    content: 'Node.js 20과 최신 브라우저를 사용해 주세요.',
-    pinned: true,
-    author_id: 'usr_inst_002',
-    created_at: '2026-04-02T12:00:00.000Z',
-  },
-  {
-    id: 'ntc_003',
-    course_id: 'crs_llm_001',
-    title: '질문 응답 시간 안내',
-    content: '질문은 댓글로 남기면 다음 실습 전에 답변드립니다.',
-    pinned: false,
-    author_id: 'usr_inst_001',
-    created_at: '2026-04-03T12:00:00.000Z',
-  },
-];
-
-export let demoLectureTranscripts: LectureTranscript[] = [];
-
-export let demoLectureNotes: LectureNote[] = [];
-
-export let demoAudioExtractions: AudioExtraction[] = [];
-
-export let demoLecturePipelines: LecturePipeline[] = [];
-
-export let demoAIUsageLogs: AIUsageLog[] = [
-  {
-    id: 'aiu_001',
-    user_id: 'usr_std_001',
-    feature: 'smart_chat',
-    provider: 'workers_ai',
-    model: 'llama-3.1-70b',
-    input_tokens: 156,
-    output_tokens: 284,
-    latency_ms: 820,
-    success: 1,
-    error_message: null,
-    created_at: '2026-04-08T08:00:00.000Z',
-  },
-  {
-    id: 'aiu_002',
-    user_id: 'usr_std_001',
-    feature: 'summary',
-    provider: 'workers_ai',
-    model: 'llama-3.1-70b',
-    input_tokens: 192,
-    output_tokens: 220,
-    latency_ms: 740,
-    success: 1,
-    error_message: null,
-    created_at: '2026-04-08T08:05:00.000Z',
-  },
-  {
-    id: 'aiu_003',
-    user_id: 'usr_std_001',
-    feature: 'quiz',
-    provider: 'workers_ai',
-    model: 'llama-3.1-70b',
-    input_tokens: 205,
-    output_tokens: 310,
-    latency_ms: 960,
-    success: 1,
-    error_message: null,
-    created_at: '2026-04-08T08:12:00.000Z',
-  },
-  {
-    id: 'aiu_004',
-    user_id: 'usr_inst_001',
-    feature: 'qa',
-    provider: 'workers_ai',
-    model: 'llama-3.1-70b',
-    input_tokens: 240,
-    output_tokens: 390,
-    latency_ms: 1120,
-    success: 1,
-    error_message: null,
-    created_at: '2026-04-08T09:00:00.000Z',
-  },
-  {
-    id: 'aiu_005',
-    user_id: 'usr_inst_001',
-    feature: 'qa',
-    provider: 'workers_ai',
-    model: 'llama-3.1-70b',
-    input_tokens: 220,
-    output_tokens: 260,
-    latency_ms: 1300,
-    success: 0,
-    error_message: 'context_not_found',
-    created_at: '2026-04-08T09:10:00.000Z',
-  },
-  {
-    id: 'aiu_006',
-    user_id: 'usr_admin_001',
-    feature: 'insights',
-    provider: 'workers_ai',
-    model: 'llama-3.1-70b',
-    input_tokens: 95,
-    output_tokens: 180,
-    latency_ms: 620,
-    success: 1,
-    error_message: null,
-    created_at: '2026-04-08T09:20:00.000Z',
-  },
-];
-
-export let demoAIIntentLogs: AIIntentLog[] = [
-  {
-    id: 'aint_001',
-    user_id: 'usr_std_001',
-    message: '이 강의 핵심만 요약해줘',
-    detected_intent: 'request_summary',
-    confidence: 0.98,
-    feature: 'summary',
-    success: true,
-    action_taken: 'summarized',
-    lecture_id: 'lec_llm_001',
-    course_id: 'crs_llm_001',
-    created_at: '2026-04-08T08:05:00.000Z',
-  },
-  {
-    id: 'aint_002',
-    user_id: 'usr_std_001',
-    message: '퀴즈 만들어줘',
-    detected_intent: 'generate_quiz',
-    confidence: 0.92,
-    feature: 'quiz',
-    success: true,
-    action_taken: 'quiz_generated',
-    lecture_id: 'lec_llm_002',
-    course_id: 'crs_llm_001',
-    created_at: '2026-04-08T08:12:00.000Z',
-  },
-  {
-    id: 'aint_003',
-    user_id: 'usr_std_001',
-    message: '이 개념을 다시 설명해줘',
-    detected_intent: 'ask_concept',
-    confidence: 0.95,
-    feature: 'smart_chat',
-    success: true,
-    action_taken: 'rag_answered',
-    lecture_id: 'lec_llm_002',
-    course_id: 'crs_llm_001',
-    created_at: '2026-04-08T08:20:00.000Z',
-  },
-  {
-    id: 'aint_004',
-    user_id: 'usr_inst_001',
-    message: '우리 강의에서 질문이 가장 많이 나온 주제가 뭐야?',
-    detected_intent: 'analyze_progress',
-    confidence: 0.88,
-    feature: 'insights',
-    success: true,
-    action_taken: 'instructor_insight',
-    lecture_id: null,
-    course_id: 'crs_llm_001',
-    created_at: '2026-04-08T09:00:00.000Z',
-  },
-  {
-    id: 'aint_005',
-    user_id: 'usr_admin_001',
-    message: '이번 주 AI 사용량을 보여줘',
-    detected_intent: 'analyze_progress',
-    confidence: 0.93,
-    feature: 'insights',
-    success: true,
-    action_taken: 'admin_overview',
-    lecture_id: null,
-    course_id: null,
-    created_at: '2026-04-08T09:20:00.000Z',
-  },
-];
-
-export let demoAIQuestionLogs: AIQuestionLog[] = [
-  {
-    id: 'aiq_001',
-    user_id: 'usr_std_001',
-    lecture_id: 'lec_llm_001',
-    course_id: 'crs_llm_001',
-    question: 'RAG는 왜 청킹이 필요한가요?',
-    answer: '청킹은 검색 단위를 잘게 나눠 관련 근거를 찾기 쉽게 해 줍니다.',
-    model: 'llama-3.1-70b',
-    success: true,
-    created_at: '2026-04-08T08:21:00.000Z',
-  },
-  {
-    id: 'aiq_002',
-    user_id: 'usr_std_001',
-    lecture_id: 'lec_llm_002',
-    course_id: 'crs_llm_001',
-    question: '요약과 질문응답을 어떻게 구분해야 하나요?',
-    answer: '요약은 원문 구조를 압축하고, 질문응답은 특정 질문에 근거를 붙여 답합니다.',
-    model: 'llama-3.1-70b',
-    success: true,
-    created_at: '2026-04-08T08:25:00.000Z',
-  },
-  {
-    id: 'aiq_003',
-    user_id: 'usr_inst_001',
-    lecture_id: 'lec_llm_001',
-    course_id: 'crs_llm_001',
-    question: 'RAG 파이프라인에서 가장 중요한 지점은?',
-    answer: '근거를 잃지 않는 검색과 재조합입니다.',
-    model: 'llama-3.1-70b',
-    success: true,
-    created_at: '2026-04-08T09:02:00.000Z',
-  },
-  {
-    id: 'aiq_004',
-    user_id: 'usr_inst_001',
-    lecture_id: 'lec_llm_002',
-    course_id: 'crs_llm_001',
-    question: '학생들이 제일 많이 묻는 개념은?',
-    answer: '청킹과 근거 검색에 대한 질문이 많습니다.',
-    model: 'llama-3.1-70b',
-    success: true,
-    created_at: '2026-04-08T09:06:00.000Z',
-  },
-];
+export * from './users';
+export * from './courses';
+export * from './media';
+export * from './ai-logs';
