+++ packages/shared/src/data/index.ts
@@ -1,1 +1,4 @@
-export * from './demo-data';
+export * from './users';
+export * from './courses';
+export * from './media';
+export * from './ai-logs';
