+++ packages/shared/src/types/ai.ts
@@ -1,0 +1,5 @@
+export * from './ai/logs';
+export * from './ai/intent';
+export * from './ai/insights';
+export * from './ai/recommendations';
+export * from './ai/chat';
