+++ packages/shared/src/types/ai/intent.ts
@@ -1,0 +1,14 @@
+export type AIIntent =
+  | 'request_summary'
+  | 'generate_quiz'
+  | 'search_content'
+  | 'ask_concept'
+  | 'ask_recommendation'
+  | 'explain_deeper'
+  | 'translate'
+  | 'compare'
+  | 'create_shortform'
+  | 'extract_audio'
+  | 'analyze_progress'
+  | 'general_chat'
+  | 'clarify';
